Fix some warnings during test execution (#10234)
## Describe your changes

Fixes some test-related warnings, e.g.:

<img width="778" alt="image"
src="https://github.com/user-attachments/assets/3c3c6177-afc7-492f-ada2-701987ccf15a"
/>

<img width="778" alt="image"
src="https://github.com/user-attachments/assets/d754b14b-9cac-4b7e-a1fc-5f39d324f612"
/>


## Testing Plan

- No logical or behavioral changes. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

+++ lib/tests/streamlit/runtime/metrics_util_test.py
@@ -117,7 +117,7 @@ class PageTelemetryTest(DeltaGeneratorTestCase):
             (datetime.date.today(), "datetime.date"),
             (datetime.datetime.today().time(), "datetime.time"),
             (pd.DataFrame(), "DataFrame"),
-            (pd.Series(), "PandasSeries"),
+            (pd.Series(dtype="float64"), "PandasSeries"),
             # Also support classes as input
             (datetime.date, "datetime.date"),
             (pd.DataFrame, "DataFrame"),
